Updating gitignore, adding main, and setup with no args

+++ source/chef/setup.rb
@@ -3,30 +3,42 @@
 # Install a full Opscode Client
 #
 
-chef_url = ARGV[0] 
-validation_client_name = nil
-unless chef_url
-  puts "The first argument to the installer must be an organization or URL"
-  puts "Example:"
-  puts "  'opscode' or 'http://chef.example.com' "
-  exit 2
-end
-if chef_url !~ /^http(s?):\/\//
-  validation_client_name = "#{chef_url}-validator"
-  chef_url = "https://api.opscode.com/organizations/#{chef_url}"
+skip_config = false
+
+case ARGV[0]
+when "upgrade"
+  skip_config = true
+when "standalone"
+  skip_config = true
 end
 
-path_to_validation_key = ARGV[1]
-unless path_to_validation_key
-  puts "The second argument ot the installer must be the path to your validation key."
-  puts "Example:"
-  puts "  '~/validation.pem'"
-  exit 3
-end
-unless File.exists?(path_to_validation_key)
-  puts "#{path_to_validation_key} does not exist!"
-  puts "Please provide a path to an existing validation key."
-  exit 4
+chef_url = ARGV[0] 
+validation_client_name = nil
+
+if skip_config
+  unless chef_url
+    puts "The first argument to the installer must be an organization or URL"
+    puts "Example:"
+    puts "  'opscode' or 'http://chef.example.com' "
+    exit 2
+  end
+  if chef_url !~ /^http(s?):\/\//
+    validation_client_name = "#{chef_url}-validator"
+    chef_url = "https://api.opscode.com/organizations/#{chef_url}"
+  end
+
+  path_to_validation_key = ARGV[1]
+  unless path_to_validation_key
+    puts "The second argument ot the installer must be the path to your validation key."
+    puts "Example:"
+    puts "  '~/validation.pem'"
+    exit 3
+  end
+  unless File.exists?(path_to_validation_key)
+    puts "#{path_to_validation_key} does not exist!"
+    puts "Please provide a path to an existing validation key."
+    exit 4
+  end
 end
 
 if Process.uid != 0
@@ -45,19 +57,21 @@ end
 
 installer_dir = File.expand_path(File.dirname(__FILE__))
 
-run_command("mkdir -p /opt/opscode")
-run_command("#{installer_dir}/embedded/bin/rsync -a --delete --exclude #{installer_dir}/setup.rb #{installer_dir}/ /opt/opscode")
-run_command("mkdir -p /etc/chef")
-run_command("cp #{path_to_validation_key} /etc/chef/validation.pem")
-run_command("chmod 600 /etc/chef/validation.pem")
-File.open("/etc/chef/client.rb", "w") do |crb|
-  crb.puts <<EOH
+if skip_config
+  run_command("mkdir -p /opt/opscode")
+  run_command("#{installer_dir}/embedded/bin/rsync -a --delete --exclude #{installer_dir}/setup.rb #{installer_dir}/ /opt/opscode")
+  run_command("mkdir -p /etc/chef")
+  run_command("cp #{path_to_validation_key} /etc/chef/validation.pem")
+  run_command("chmod 600 /etc/chef/validation.pem")
+  File.open("/etc/chef/client.rb", "w") do |crb|
+    crb.puts <<EOH
 log_level                :info
 log_location             STDOUT
 chef_server_url          "#{chef_url}"
 EOH
-  if validation_client_name
-    crb.puts "validation_client_name   '#{validation_client_name}'"
+    if validation_client_name
+      crb.puts "validation_client_name   '#{validation_client_name}'"
+    end
   end
 end
 
@@ -66,5 +80,6 @@ Thanks for installing Chef!
 
 You can now run /opt/opscode/bin/chef-client to configure your system.
 EOH
+
 exit 0
